init.rc: Fix compcache and cleanup init file
The "import" keyword is only parsed once, for a one shot execution, during
the initial section setup, and before running "on fs". Having an import of
a file that's located in a filesystem other than root will result in an error
like

<3> init: could not import file '/system/etc/init.local.rc' from '/init.rc'

So... any files imported into init need to be moved to the root fs.
While we're at it, move init.rc changes that are specific to CM (and don't
involve modification of preexisting configs) into this file, to ease future
upstream merges.

Change-Id: I0f70ffad874648198acde67c12e1a7833552ad7c

Conflicts:
	rootdir/init.rc